requirements/_base.txt

https://www.djangoproject.com/download/1.5c1/tarball/
bpython==0.12
django-braces==0.2.1
django-model-utils==1.1.0
South==0.7.6
pinax-theme-foundation
psycopg2==2.4.6
django-extensions==1.0.3
python-dateutil==2.1
django-countries==1.5
django-admin-tools==0.4.1
